@@ -19,7 +19,7 @@
         <maven.build.timestamp.format>yyyy-MM-dd HH:mm</maven.build.timestamp.format>
 
         <google.gson.version>2.8.5</google.gson.version>
-        <javafx.version>11</javafx.version>
+        <javafx.version>14</javafx.version>
 
         <junit5.surefire.version>1.3.1</junit5.surefire.version>
         <junit5.version>5.3.1</junit5.version>
@@ -135,8 +135,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.8.0</version>
                 <configuration>
-                    <source>11</source>
-                    <target>11</target>
+                    <source>14</source>
+                    <target>14</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -181,22 +181,6 @@
                 <artifactId>maven-resources-plugin</artifactId>
                 <version>3.0.2</version>
                 <executions>
-                    <execution>
-                        <id>copy-resources-embed</id>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>copy-resources</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>${basedir}/target/jfx/app/embedded</outputDirectory>
-                            <resources>
-                                <resource>
-                                    <directory>${project.basedir}/../embedded</directory>
-                                    <filtering>false</filtering>
-                                </resource>
-                            </resources>
-                        </configuration>
-                    </execution>
                     <execution>
                         <id>copy-resources-logging</id>
                         <phase>validate</phase>
