@@ -6,6 +6,7 @@
 
 package com.thecoderscorner.menu.editorui.generator.plugin;
 
+import java.nio.file.Path;
 import java.util.List;
 
 public class CodePluginConfig {
@@ -17,6 +18,7 @@ public class CodePluginConfig {
     private String vendor;
     private String vendorUrl;
     private List<CodePluginItem> plugins;
+    private Path path;
 
     public CodePluginConfig() {
         // for reflection / serialisation
@@ -37,33 +39,72 @@ public class CodePluginConfig {
         return moduleName;
     }
 
+    public void setModuleName(String moduleName) {
+        this.moduleName = moduleName;
+    }
+
     public String getName() {
         return name;
     }
 
+    public void setName(String name) {
+        this.name = name;
+    }
 
     public String getVersion() {
         return version;
     }
 
-    public List<CodePluginItem> getPlugins() {
-        return plugins;
+    public void setVersion(String version) {
+        this.version = version;
     }
 
     public String getLicense() {
         return license;
     }
 
+    public void setLicense(String license) {
+        this.license = license;
+    }
+
     public String getLicenseUrl() {
         return licenseUrl;
+    }
+
+    public void setLicenseUrl(String licenseUrl) {
+        this.licenseUrl = licenseUrl;
     }
 
     public String getVendor() {
         return vendor;
     }
 
+    public void setVendor(String vendor) {
+        this.vendor = vendor;
+    }
+
     public String getVendorUrl() {
         return vendorUrl;
+    }
+
+    public void setVendorUrl(String vendorUrl) {
+        this.vendorUrl = vendorUrl;
+    }
+
+    public List<CodePluginItem> getPlugins() {
+        return plugins;
+    }
+
+    public void setPlugins(List<CodePluginItem> plugins) {
+        this.plugins = plugins;
+    }
+
+    public Path getPath() {
+        return path;
+    }
+
+    public void setPath(Path path) {
+        this.path = path;
     }
 
     @Override
