@@ -10,10 +10,10 @@ import com.thecoderscorner.menu.domain.EditableTextMenuItemBuilder;
 import com.thecoderscorner.menu.domain.MenuItem;
 import com.thecoderscorner.menu.domain.state.MenuTree;
 import com.thecoderscorner.menu.editorui.generator.CodeGeneratorOptions;
+import com.thecoderscorner.menu.editorui.generator.core.CreatorProperty;
+import com.thecoderscorner.menu.editorui.generator.core.NameAndKey;
+import com.thecoderscorner.menu.editorui.generator.plugin.*;
 import com.thecoderscorner.menu.editorui.generator.util.LibraryStatus;
-import com.thecoderscorner.menu.pluginapi.*;
-import com.thecoderscorner.menu.pluginapi.model.CodeVariableBuilder;
-import com.thecoderscorner.menu.pluginapi.model.FunctionCallBuilder;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -23,30 +23,43 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.*;
+import java.util.Comparator;
+import java.util.List;
+import java.util.UUID;
 import java.util.function.Consumer;
 
+import static com.thecoderscorner.menu.editorui.generator.plugin.EmbeddedPlatform.ARDUINO32;
+import static com.thecoderscorner.menu.editorui.generator.plugin.EmbeddedPlatform.ARDUINO_AVR;
 import static com.thecoderscorner.menu.editorui.util.TestUtils.assertEqualsIgnoringCRLF;
 import static com.thecoderscorner.menu.editorui.util.TestUtils.buildSimpleTreeReadOnly;
-import static com.thecoderscorner.menu.pluginapi.EmbeddedPlatform.ARDUINO32;
-import static com.thecoderscorner.menu.pluginapi.EmbeddedPlatform.ARDUINO_AVR;
-import static com.thecoderscorner.menu.pluginapi.PluginFileDependency.PackagingType;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.*;
 
 public class ArduinoGeneratorTest {
 
-    private Path dir;
+    private Path projectDir;
+    private Path pluginDir;
+    private Path rootDir;
+    private CodePluginConfig pluginConfig;
 
     @BeforeEach
     public void setUp() throws Exception {
-        dir = Files.createTempDirectory("tcmenu");
+        rootDir = Files.createTempDirectory("tcmenutest");
+        projectDir = rootDir.resolve("project");
+        Files.createDirectories(projectDir);
+
+        pluginDir = rootDir.resolve("plugin");
+        pluginDir = DefaultXmlPluginLoaderTest.makeStandardPluginInPath(pluginDir);
+        var embeddedPlatforms = new PluginEmbeddedPlatformsImpl();
+        var loader = new DefaultXmlPluginLoader(embeddedPlatforms);
+        pluginConfig = loader.loadPluginLib(pluginDir);
+
     }
 
     @AfterEach
     public void tearDown() throws Exception {
-        Files.walk(dir)
+        Files.walk(rootDir)
                 .sorted(Comparator.reverseOrder())
                 .map(Path::toFile)
                 .forEach(File::delete);
@@ -80,28 +93,34 @@ public class ArduinoGeneratorTest {
         ArduinoLibraryInstaller installer = Mockito.mock(ArduinoLibraryInstaller.class);
         Mockito.when(installer.statusOfAllLibraries()).thenReturn(new LibraryStatus(true, true, true));
 
-        List<EmbeddedCodeCreator> generators = unitTestGenerator();
         CodeGeneratorOptions standardOptions = new CodeGeneratorOptions(
                 ARDUINO32.getBoardId(),
                 "", "", "",
                 List.<CreatorProperty>of(),
                 UUID.randomUUID(),
                 "app",
-                recursiveName);
+                recursiveName, false);
         ArduinoGenerator generator = new ArduinoGenerator(adjuster, installer, platform, standardOptions);
 
-        assertTrue(generator.startConversion(dir, generators, tree, new NameAndKey("uuid1", "tester"), List.of()));
+        var firstPlugin = pluginConfig.getPlugins().get(0);
+        firstPlugin.getProperties().stream()
+                .filter(p -> p.getName().equals("SWITCH_IODEVICE"))
+                .findFirst()
+                .ifPresent(p -> p.getProperty().setValue("io23017"));
+
+        assertTrue(generator.startConversion(projectDir, pluginConfig.getPlugins(), tree,
+                new NameAndKey("uuid1", "tester"), List.of(), false));
 
         assertEquals("GenState", generator.makeNameToVar(generateItemWithName("Gen &^%State")));
         assertEquals("ChannelÖôóò", generator.makeNameToVar(generateItemWithName("ChannelÖôóò")));
 
-        var cppGenerated = new String(Files.readAllBytes(dir.resolve(dir.getFileName() + "_menu.cpp")));
-        var hGenerated = new String(Files.readAllBytes(dir.resolve(dir.getFileName() + "_menu.h")));
-        var pluginGenerated = new String(Files.readAllBytes(dir.resolve("replacementSource.h")));
+        var cppGenerated = new String(Files.readAllBytes(projectDir.resolve(projectDir.getFileName() + "_menu.cpp")));
+        var hGenerated = new String(Files.readAllBytes(projectDir.resolve(projectDir.getFileName() + "_menu.h")));
+        var pluginGeneratedH = new String(Files.readAllBytes(projectDir.resolve("source.h")));
+        var pluginGeneratedCPP = new String(Files.readAllBytes(projectDir.resolve("source.cpp")));
 
         var cppTemplate = new String(getClass().getResourceAsStream(templateToUse + ".cpp").readAllBytes());
         var hTemplate = new String(getClass().getResourceAsStream(templateToUse + ".h").readAllBytes());
-        var expectedPlugin = new String(getClass().getResourceAsStream("/generator/replacementExpected.h").readAllBytes());
 
         cppGenerated = cppGenerated.replaceAll("#include \"tcmenu[^\"]*\"", "replacedInclude");
         cppTemplate = cppTemplate.replaceAll("#include \"tcmenu[^\"]*\"", "replacedInclude");
@@ -110,40 +129,11 @@ public class ArduinoGeneratorTest {
         // then make sure the change is good before adjusting the templates.
         assertEqualsIgnoringCRLF(cppTemplate, cppGenerated);
         assertEqualsIgnoringCRLF(hTemplate, hGenerated);
-        assertEqualsIgnoringCRLF(expectedPlugin, pluginGenerated);
+        assertEqualsIgnoringCRLF("CPP_FILE_CONTENT 10 otherKey", pluginGeneratedCPP);
+        assertEqualsIgnoringCRLF("H_FILE_CONTENT 10 otherKey", pluginGeneratedH);
 
         Mockito.verify(adjuster).makeAdjustments(any(Consumer.class),
-                eq(dir.resolve(dir.resolve(dir.getFileName() + ".ino")).toString()),
-                eq(dir.getFileName().toString()), anyCollection());
-    }
-
-    private List<EmbeddedCodeCreator> unitTestGenerator() {
-        EmbeddedCodeCreator gen = new AbstractCodeCreator() {
-            @Override
-            public List<CreatorProperty> properties() {
-                return List.of(new CreatorProperty("A_DEFINE", "blah", "2", SubSystem.INPUT));
-            }
-
-            @Override
-            protected void initCreator(String root) {
-                addVariable(new CodeVariableBuilder()
-                        .requiresHeader("header1.h", false)
-                        .variableName("varName").variableType("VarType")
-                        .param("1234.34")
-                        .exportNeeded()
-                );
-
-                addFunctionCall(new FunctionCallBuilder().functionName("begin").objectName("lcd").param(16).param(2));
-
-                addLibraryFiles(new PluginFileDependency(
-                        "testSrc/replacementSource.h",
-                        PackagingType.WITH_PLUGIN,
-                        Map.of(
-                                "Replacement1.h", "ChangedHeader.h",
-                                "ReplacementServer", "ChangedServer")
-                ));
-            }
-        };
-        return Collections.singletonList(gen);
+                eq(projectDir.resolve(projectDir.resolve(projectDir.getFileName() + ".ino")).toString()),
+                eq(projectDir.getFileName().toString()), anyCollection());
     }
 }
