@@ -9,8 +9,8 @@ package com.thecoderscorner.menu.editorui.project;
 import com.thecoderscorner.menu.domain.MenuItem;
 import com.thecoderscorner.menu.domain.state.MenuTree;
 import com.thecoderscorner.menu.editorui.generator.CodeGeneratorOptions;
+import com.thecoderscorner.menu.editorui.generator.core.CreatorProperty;
 import com.thecoderscorner.menu.editorui.util.TestUtils;
-import com.thecoderscorner.menu.pluginapi.CreatorProperty;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -21,8 +21,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.*;
 
-import static com.thecoderscorner.menu.pluginapi.EmbeddedPlatform.ARDUINO_AVR;
-import static com.thecoderscorner.menu.pluginapi.SubSystem.DISPLAY;
+import static com.thecoderscorner.menu.editorui.generator.core.SubSystem.DISPLAY;
+import static com.thecoderscorner.menu.editorui.generator.plugin.EmbeddedPlatform.ARDUINO_AVR;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class FileBasedProjectPersistorTest {
@@ -55,7 +55,7 @@ public class FileBasedProjectPersistorTest {
                 "uuid2",
                 "uuid3",
                 Collections.singletonList(new CreatorProperty("name", "desc", "123", DISPLAY)),
-                APPLICATION_UUID, "app name", false
+                APPLICATION_UUID, "app name", false, false
         );
         persistor.save(projFile.toString(), tree, options);
 
